moved the ArgsForm component to it's own module.
